Remove mom6_index_reverse in favor of ocean_side; fix land-cap fixtures; consolidate Segment tests
Segment.from_hgrid/from_lonlat now take ocean_side as the sole,
required way to specify which side of a line is open ocean --
mom6_index_reverse and its mutual-exclusivity dance are gone.
Segment.cardinal derives ocean_side internally from the opposite
compass direction; to_spec/from_spec now round-trip ocean_side.

Also fixes the sketch-domain and regrid-tracers test fixtures, which
had gone stale against the land-capped-endpoints check added in
720364d (segment_002/004/010 ended in open water with no land behind
their endpoints); widens headland A and narrows segment_004's span so
every fixture is genuinely land-capped, with hardcoded position-string
assertions updated to match.

Consolidates all Segment-focused tests (test_segment_multi_boundary_domain.py,
test_detect_open_cardinal_boundaries.py, and the Segment construction/
position-string/regrid_velocity_tracers/regrid_tides tests previously
in test_regridding.py) into a single tests/test_segment.py, leaving
test_regridding.py to genuinely regridding.py-module-level tests. Adds
previously-missing coverage for to_spec/from_spec round-tripping and
from_lonlat.

Adds demos/segment-basics-demo.ipynb, a short standalone tour of
building OBC segments via Segment alone.
